Rezept/backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
sqlalchemy<2.0.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
psycopg2-binary==2.9.1
celery==5.1.2
redis==3.5.3
gunicorn==20.1.0
python-dotenv==0.19.0
pdfplumber==0.6.0
Pillow>=8.4
pytesseract==0.3.8
numpy==1.21.2
pandas==1.3.3
werkzeug==2.0.1
requests==2.26.0